Create the libraries folder on demand in install_libraries_from_path()
It is necessary to make the parent folder in install_platforms_from_path(), since the folder will be different for every platform, so it makes sense to mirror that behavior in the equivalent libraries function.

The other installation methods already will create the parent folder if it doesn't already exist, so it's only necessary to add this to install_libraries_from_path() after removing it from install_libraries().