Allow installation of platform dependencies from local path
This change allows the action to be used for easy compilation testing of platforms.

The easiest way to install the platform's tools dependencies is via Board Manager. If the platform from the local path was also installed via Board Manager (as identified by them having the same name value, the Board Manager platform will be replaced by the one from the local path.