Rename size-deltas-report-folder-name input to sketches-report-path
Now that the report is always generated and is used to transfer data to the reportsizetrends module, it's misleading for the input name to reference "deltas".

The previous input name is still supported. It is now undocumented and a warning will be printed in the workflow log when it is used.

Although no longer an accurate name, the previous default value is retained to avoid breaking workflows that use the default input value and then reference that path in the actions/upload-artifact step.