Support ~ in input paths
The Arduino user folder (sketchbook) is at /home/github/Arduino and the Arduino data folder at /home/github, but the /home/github path is not obvious or documented and there is no way to use the HOME environment variable in inputs to this action. So the most user friendly way to specify these paths outside the workspace is using ~. This is not supported by default.