Use consistent approach to processing user-provided paths
Paths from user input are converted to absolute paths as early as possible, then those absolute paths are used from there on. If possible, the path is converted to a user-friendly path relative to the root of the repository (GITHUB_WORKSPACE) before output to the log or sketches report.

This is now more important due to the addition of support for ~, which requires processing through the absolute_path() function to work.