Update sketches report key name to reflect expanded scope
The sketches report now contains data for all compiled sketches, rather than a single one. The "sketch" key was previously used for the sketch path, but it is now used to contain all the data for the individual sketches. The move to reporting data for all sketches already causes a breaking change to the report, so updating the key name does no additional harm.