Change sketches report key from "fqbn" to "board"
Arduino CLI's issue tracker has shown that users have a lot of trouble with FQBN. For this reason, Arduino CLI will soon support the use of "board aliases" in place of FQBN. This is likely to become more commonly used than the FQBN. For this reason, it makes sense to change the key name to match the expected usage. "board" still applies well even if the value is an FQBN, whereas the reverse is not so.

The move to reporting data for all sketches already causes a breaking change to the report, so updating the key name does no additional harm.