Report relative memory usage data
Previously, only absolute memory usage data was reported.

The memory usage as a percentage of the total available may often be even even more useful. For example, an increase of 10 bytes is much more significant on an ATtiny85 than it is on an ATSAMD21G18.

In order to calculate relative memory usage, it's necessary to know the total available memory. Total available memory is determined in the same manner as the absolute memory usage: by a regular expression on the compilation output. Some platforms are not configured to provide this information. In that case, the maximum and relative memory usage data will be recorded as "N/A".