Overwrite previous installations of dependencies
Library dependencies of Library Manager sourced libraries (as defined in its library.properties metadata file) are automatically installed by Arduino CLI.

Libraries from Library Manager sources are installed before the libraries from other sources. When the user of the action defined a library dependency that was also automatically installed via the Library Manager dependency system, the workflow step using the action would fail on the installation of the library from the user's definition due to an error caused by the library being already installed.

My first instinct for how to fix the failure caused by conflict with an existing installation was to simply change the order the library dependency sources are installed so the Library Manager sourced libraries are installed last. However, there is another factor: If a library with a version lower than the latest version available from Library Manager is already installed, Arduino CLI updates it with the version from Library Manager. The exact version of the library dependencies specified by the user of the script must be retained, so this means the reordering solution won't work.

The solution is to leave the installation order so the Library Manager sourced libraries are installed first, then configure the installation of libraries from the other sources to overwrite existing installations.

Adding this "force" capability to the `compilesketches.CompileSketches.install_from_path()` function also allowed moving the overwrite code from from its previous unlikely home in  `compilesketches.CompileSketches.get_platform_installation_path()`.