Install each Library Manager sourced library separately
`arduino-cli lib install` fails if one of the libraries in the list has a dependency on another, but an earlier version of the dependency is specified in the list.

For example, the following command doesn't install any libraries and returns exit status 1:

$ arduino-cli lib install Arduino_ConnectionHandler@0.4.8 ArduinoIoTCloud
Arduino_ConnectionHandler@0.4.8 depends on Arduino_DebugUtils@1.1.0
Arduino_ConnectionHandler@0.4.8 depends on WiFi101@0.16.0
Arduino_ConnectionHandler@0.4.8 depends on WiFiNINA@1.7.1
Arduino_ConnectionHandler@0.4.8 depends on MKRGSM@1.5.0
Arduino_ConnectionHandler@0.4.8 depends on MKRNB@1.3.2
Arduino_ConnectionHandler@0.4.8 depends on MKRWAN@1.0.12
Arduino_ConnectionHandler@0.4.8 depends on Arduino_ConnectionHandler@0.4.8
ArduinoIoTCloud depends on Arduino_DebugUtils@1.1.0
ArduinoIoTCloud depends on ArduinoMqttClient@0.1.5
ArduinoIoTCloud depends on ArduinoECCX08@1.3.4
ArduinoIoTCloud depends on WiFiNINA@1.7.1
ArduinoIoTCloud depends on MKRGSM@1.5.0
ArduinoIoTCloud depends on ArduinoIoTCloud@0.11.1
ArduinoIoTCloud depends on RTCZero@1.6.0
ArduinoIoTCloud depends on WiFi101@0.16.0
ArduinoIoTCloud depends on MKRNB@1.3.2
ArduinoIoTCloud depends on MKRWAN@1.0.12
ArduinoIoTCloud depends on Arduino_ConnectionHandler@0.4.9
The library Arduino_ConnectionHandler is required in two different versions: 0.4.9 and 0.4.8

The solution is to install one library at a time (even though `arduino-cli lib install` supports installing multiple libraries at once). This also allows the user to control which version of the library is installed in the end by the order of the list passed via the libraries input.