Rename all occurrences of "size deltas report" to "deltas report"
The report may contain deltas data other than size, so the previous name was too constraining.

The only remaining use of "size deltas report" is the default sketches report path: 'size-deltas-reports'.

For backwards compatibility, the `enable-size-deltas-report` input will still be supported, but a warning will be displayed in the workflow run log notifying the user of the name change.