Split repository's CI workflow into functional parts.
In the experimental repository the action was migrated from, the system was to have a single workflow for each action hosted in the repository.

Although it is a sensible approach in that context, it meant combining things that are only loosely related into a single workflow. Now that the action has a dedicated repository, it makes sense to split it into multiple workflows.