Remove unnecessary conditional from example workflow snippet
Now that arduino/compile-sketches always generates a sketches report, the conditional in the workflow snippet in the readme is unnecessary. Although a user may decide they want to do advanced configuration of their workflow to control when the artifact is generated, there is no requirement that it only be made on pull_request events, so this only adds unnecessary complexity to the example.

The example should contain only the minimum necessary to effectively demonstrate the basic use of the action.