Use appropriate default value for sketches-report-path input
In the time since it was first conceived, the use of the report file created by the compile-sketches action has been expanded from only being a way to pass memory usage change data to the report-size-deltas action to a general purpose report of data generated from the sketch compilations, including:

- Non-deltas size data consumed by arduino/report-size-trends
- Compilation status for each sketch/board combination
- Compiler warning count

It's possible additional information will be added to the report over time.

For this reason, the previous default `size-deltas-reports` was no longer appropriate and might be the source of confusion.

Unfortunately, this is a breaking change for people who:

- Did not specify a value to the sketches-report-path input and,
- have an additional step in their workflow (e.g., actions/upload-artifact) that relies on the sketches reports being located under the folder resulting from the previous default