Give Python virtual environment access to system site-packages directory
Use the` venv --system-site-packages` flag when creating the Python virtual environment the action's script runs in to allow users to install Python package dependencies of their sketch's boards platform dependencies.

The immediate need for this is that the ESP32 boards platform has a dependency on the `pyserial` Python package. The alternative would be to always install this specific package in the setup script, but allowing the user to install arbitrary Python package dependencies makes the action more flexible.

The cons:
- Increased complexity of workflows compiling for ESP32 platform boards.
- Risk of the runner's default packages interfering with the script.