Fix CI workflow for Python linting
The workflow was broken by the Python version it was using no longer being available from the actions/setup-python
action.

Rather than continuing in this fashion, it seems better to use the same setup script as is used by the action and unit
tests. This will ensure everything is done under the same environment.