Update integration test to use cli-compile-flags input
A sketch was added that is designed to fail compilation if the `arduino-cli compile` command does not contain the flags
specified via the input. Because the `--build-property` flag used in the test was not present in Arduino CLI 0.13.0 (the
equivalent flag was `--build-properties` at that time), it was necessary to update the Arduino CLI version used by the
action in the test.