Support new Arduino CLI JSON output keys
A breaking change was made to the key names of Arduino CLI's JSON output in its 0.18.0 release. This action was dependent
on some of the keys that changed.

Because the action has a `cli-version` input that allows it to be used with any Arduino CLI version, it is necessary to
support both interface styles. Even though there is not currently extensive use of the interface, it seemed best to set
up a translation system to support all usage of the interface rather than just doing a spot fix at the site of the
current breakage.