Correctly detect and handle installation path collisions
`compilesketches.CompileSketches.install_from_path()` has code for detecting a pre-existing file/folder at the
installation destination path and either removing it or else failing with a helpful error message depending on the
setting of the `force` parameter.

The previous detection code did not work correctly in the event the pre-existing item was a broken symlink.

The previous removal code did not work correctly in the event the pre-existing item was a symlink or file.