Clean up dependencies installation symlinks on exit
The code used to install Arduino CLI, platform, and library dependencies does this by creating symlinks in the
appropriate locations on the runner. The files the script gathers via clone or download, are stored in a temporary folder
that is deleted when the script exits. This means the installation symlink for those files has no possible use after the
script finishes.

This particular pollution of the runner's environment is resolved by also registering a call to delete on exit as each
symlink is created.