Use `actions/setup-python` action to install Python for action
At the time the arduino/compile-sketches action was converted to a composite action, the only step type supported for
this action type was "run", where shell commands are executed. For this reason, it was necessary to use shell commands
to install Python.

Since that time, support was added for using other actions in composite action steps. This allows Python to be set up
for use by the action in a more simple and efficient manner via the `actions/setup-python` action.

The Python-based infrastructure workflows are also migrated to using `actions/setup-python` action for installing
Python. In order to ensure the same version of Python is used for CI as the action, the version of Python used by the
project is defined in a `.python-version` file, and all `actions/setup-python` steps pointed to that file.

In order to ensure the stability of the action, the `actions/setup-python` action is pinned to a patch version. Although
the common practice is to only pin the major version of actions in CI workflows, in this case there is no benefit to
doing so for the `actions/setup-python` action since patch level bumps are going to be received from Dependabot
regardless.