Remove leftover items in workflow path filter
Previously, a script was used to install Python and venv for use by the action. The path of this script was added to the
paths filter in the relevant workflows so that they would be triggered on any change to the script.

The script was removed in favor of the use of the actions/setup-python script for installation of Python. The script
path was not removed from the workflow at that time.