Restore support for external platform Python package dependencies
Although Arduino boards platforms traditionally have bundled all dependencies, there are some cases where the user will
need to use a preceding workflow step to install external dependencies in the GitHub Actions runner environment.

Since it is a "composite action", this is feasible and was previously supported. However, support for Python package
dependencies is more complicated due to the action script running in a Python virtual environment which isolates it from
packages installed globally. The workaround for this was to use the `--system-site-packages` flag in the venv invocation
that created the action's Python virtual environment. This approach introduced a requirement that the project Python
version match the minor version series of the runner system Python. That requirement was violated when the workflow's
runner was updated to `ubuntu-22.04` which has system Python 3.10.x, whereas the action used 3.8. The breakage went
unnoticed because the sole Python package platform dependency in real world usage, pyserial, had been preinstalled in
the runner's Linux package manager "site-packages" folder, which caused it to be installed in the action's virtual
environment even though it would not have if depending on the user installation of the package as was previously the
case.

The solution is to use a path configuration file to add the path of the system Python's "user site-packages" folder
(where package dependencies installed via the workflow are located) to the module search paths used by the action's
Python virtual environment.

This is actually superior to the previous approach of using the `--system-site-packages` flag in the venv invocation
(this flag is also supported by pipx and Poetry) because:

- It avoids the chance of interference with the virtual environment through the unnecessary introduction of the global
  "site-packages" path into the module search paths along with the intended introduction of the "user site-packages"
  path.
- It allows any version of Python to be used to run the action, rather than being forced to use the same minor version
  series as the runner's system Python (which can be updated at any time and also depends on the workflow configuration)