Test support of pyserial external dependency of ESP32 platform
Although Arduino boards platforms traditionally have bundled all dependencies, the ESP32 boards platform expects the
pyserial Python package to be preinstalled on Linux systems.

Although there is already test coverage and support for user installation of arbitrary Python package dependencies of a
platform, the situation is more complex when it comes to this important boards platform. The problem is that there is an
established use pattern that should be supported to avoid breaking user workflows.

The added test checks for regressions in that support.