Group action setup log output
The setup of the action produces a significant amount of output, which is printed in the worflow run logs. This output
is not of interest to the user so long as the action itself is working correctly so it is only noise in their logs.

A log group was previously used to collapse all that setup output in the logs. This was inadvertently removed during
recent development work. It is hereby restored.