Update infrastructure for spell checking project contents
Arduino tooling projects use a standardized infrastructure. A centralized collection of reusable infrastructure assets
is maintained in a dedicated repository:

https://github.com/arduino/tooling-project-assets

Since the time this project's infrastructure was installed, some advancements have been made in the upstream "template"
assets. The project's infrastructure is hereby brought up to date with the state of the art upstream assets.

The significant changes:

- Migration to the use of the Task task runner to allow contributors to easily run the same operations locally as is
  done by the CI system on GitHub.
- Use of Poetry and Dependabot for controlled management of the dependencies.