Update infrastructure for GitHub issue label management
Arduino tooling projects use a standardized infrastructure. A centralized collection of reusable infrastructure assets
is maintained in a dedicated repository:

https://github.com/arduino/tooling-project-assets

Since the time this project's infrastructure was installed, some advancements have been made in the upstream "template"
assets. The project's infrastructure is hereby brought up to date with the state of the art upstream assets.

The significant changes:

- Use of npm and Dependabot for controlled management of the dependencies.