Fix Dependabot target path for Python dependencies
When the project was migrated to using Poetry as the dependencies management system, the dependencies metadata file was
moved from the Python project subfolder to the root of the repository, where it can be used to manage all development
tool dependencies in addition to the Python code dependencies.

The Dependabot configuration file was not updated accordingly at that time.