Resolve conflict between Black and Flake8 code styles
The project uses the Arduino tooling standard target line length of 120.

Flake8's rule E501 uses this as a hard limit, erroring if any line in a Python file exceeds it.

Black uses 120 as a target length, but may format some lines longer than 120 when it considers complying with the limit
to be harmful.

So rule E501 must be disabled. The line length limit will still be enforced by the formatting check.