Add infrastructure for maintaining the table of contents
A task and GitHub Actions workflow are provided here for for maintaining the the table of contents in the project's
readme.

The `markdown:toc` task automatically updates the table of contents to match the contents of the readme.

On every push or pull request that affects the repository's Markdown files that contain a generated table of contents,
the GitHub Actions workflow checks whether the table of contents matches the file structure.