Always add size summary report to sketches report
In addition to per-sketch, per-board memory usage data, the sketches report contains a per-board summary of the results
from all the compilations for that board.

Previously, when deltas data was not available, this summary section was arbitrarily omitted entirely from the sketches
report. However, there is still board-universal information even without deltas:

- Memory types
- Maximums for each memory type

It is useful to have that information available on a per-board level so a summary report should always be added to the
sketches report.