Document how to source platform/library dependencies from private repo
The action supports installing platform and library dependencies of the sketches by cloning the repository specified via
the source-url field of the platforms or libraries inputs.

The information provided in the relevant documentation in the readme should be sufficient for the action user to
accomplish this when the dependency is in a public repository. Although that information is also applicable for private
repositories, it is more complex to make the necessary configuration of the workflow, repository, and user account. So
the users might appreciate some additional guidance for this use case.

Since this configuration is general to using Git, GitHub, and GitHub Actions workflows rather than something specific to
the action itself, it is out of scope for the primary documentation in the readme. An item in the supplemental FAQ will
be the most appropriate place for this information.