Handle Arduino CLI 1.x `core list` data format
The action parses the output of the `arduino-cli core list --format json` command.

There was a breaking change to the data format of that output was changed in the 1.0.0 release of Arduino CLI.

The action code is updated to handle the new output, while also retaining backwards compatibility with the pre-1.x data
format in order to support workflows that pin the Arduino CLI version to >1.0.0 via the action's "cli-version" input.