Remove unused Arduino CLI output key translations
The action parses the JSON formatted output of the `arduino-cli core list` command. The action depends on the output
having a specific data format, which includes the key names. If the Arduino CLI developers change the name of a key used
by the action, it breaks the action.

Some time ago, there was a comprehensive review and standardization of the Arduino CLI output key names, which resulted
in many changes to the key names. A key name translation system was added to the action code in order to handle that
breaking change. Although the action's current code only relies on a couple of keys from a single command, it is planned
to use other parts of the output over time (e.g., getting the compilation data now available from the `arduino-cli
compile --format json` output instead of parsing the text output as was necessary at the time the action was created).
There was an impression of finality to the new key names. For these reasons, the decision was made to implement a
comprehensive translation system rather than limiting the scope to only the keys used by the action at the time.

Since that time, various key name changes have occurred at various times. Since none of the changed keys were used by
the action, this did not cause breakage, but it did render the comprehensive key name translation system outdated. A key
used by the action has been renamed once again in the Arduino CLI 1.0.0 release, which requires the translation system
to be redesigned and updated. The fact that the comprehensive system has not been kept updated indicates that it is not
feasible to maintain such a system. So, rather than investing the significant amount of effort that would be required to
review all the changes made to the Arduino CLI output key names and update the comprehensive translation system
accordingly, the translation system is trimmed back to only cover the keys actually used by the action. It will be
expanded on demand.