Add command output key translation for Arduino CLI 1.0.0
The action parses the output of the `arduino-cli core list --format json` command.

The name of one of the keys in that output data used by the action was changed in the 1.0.0 release of Arduino CLI,
breaking the action when used with that version of Arduino CLI (which is the default behavior).

The action code base already has a key name translation system, which was implemented to handle breakage caused by a
previous round of key name changes. However, that system was designed under the assumption that the key names would be
stable following the comprehensive redesign that caused the previous changes. It was not designed to handle multiple
changes to the name of a given key as has now happened. For this reason, the translation system has been redesigned to
support any number of key name changes that might occur.