Configure PYTHONPATH via pytest configuration file
Previously, the python:test task set PYTHONPATH via environment variable. That approach is fine when running the tests
via the task, but meant that PYTHONPATH was not correctly configured when running the tests via the interface of an IDE.
Each contributor would be forced to configure their IDE in order to allow it to run the tests successfully.

Configuring PYTHONPATH via the pytest.ini file makes this universal.