Use pyproject.toml as Python version source
All execution, development and validation for the project is performed using a specific version of Python.

Poetry is used for Python package dependencies management. Poetry installs dependencies into a virtual environment. For
this reason, the Poetry configuration includes a Python version number, which must match the correct version of Python
for the project. This configuration is stored in the `pyproject.toml` file.

Python is installed in the GitHub Actions runner environments using the actions/setup-python action, which also must be
configured to install the correct version of Python. Previously a the version number for use by the actions/setup-python
action was defined in the `.python-version` file. This allowed the Python version for use by the
arduino/compile-sketches action's execution and all the workflows that make up its infrastructure to be configured from
a single source. However, Poetry does not support the `.python-version` file so we were stuck with two copies of the
Python version information, which must be kept in sync.

Fortunately, support for using `pyproject.toml` as the version file for the actions/setup-python action was recently
added. This means it is now possible for all components of the project infrastructure to get the Python version from a
single source.