Install Poetry in dependent tasks
The project's Python package dependencies are managed by the Poetry tool.

In addition to the packages used by the action script, project development tools from Python packages are also managed
by Poetry. For this reason, Poetry is a dependency of several of the tasks in the taskfile.

Previously, there was an undocumented requirement that the contributor have Poetry installed and in their system path.
The reason for that approach was that, at the time, the practice was to make a global system installation of Poetry,
which might cause problems for some contributors. Since that time, the project has migrated to using pipx to install
Poetry. This tool installs Poetry in an virtual environment.

Of course, even though this means manual installation of Poetry is no longer the responsibility of the contributor,
installation of pipx is. However, the officially recommended mechanism for installing Poetry is now pipx so the
contributor would have likely ended up manually installing pipx anyway as a prerequisite for the Poetry installation
procedure.