Don't upload multiple times to same artifact in integration test workflow
The `all-inputs` job of the "Run integration tests" GitHub Actions workflow is configured to generate multiple parallel
jobs. A GitHub Actions workflow artifact was used to transfer the sketches report files generated by the
"arduino/compile-sketches" action in each job to the subsequent job. The "actions/upload-artifact" and
"actions/download-artifact" actions are used for this purpose.

Previously, a single artifact was used for the transfer of all the sketches reports, with each of the parallel jobs
uploading its own generated files to that single artifact. However, support for uploading multiple times to a single
artifact was dropped in version 4.0.0 of the "actions/upload-artifact" action. So it is now necessary to use a dedicated
artifact for each of the parallel jobs. These artifacts  can be downloaded in aggregate by using the artifact name
globbing and merging features which were introduced in version 4.1.0 of the "actions/download-artifact" action.