Update poetry:sync task for Poetry 2.0.0 CLI
Previously, it was necessary to add the `--no-update` flag to the `poetry lock` command in order to prevent it from
updating the dependency versions in the lockfile. The behavior of the command has now been changed to always behave that
way, and so the `--no-update` flag was removed from the command.