Configure Poetry installation command to install project-defined version even when already installed
One of the reasons for adding the Poetry installation infrastructure to the project was to control the version of Poetry
that is used by contributors and the infrastructure.

The default behavior of the `pipx install` command is to skip the installation if Poetry is already installed, even
though a different version is installed than that specified in the package version constraint, not even throwing an
error exit status to communicate that the installation was not performed.

It is possible to cause pipx to have the intended behavior of always installing the specified version by adding the
`--force` flag to the command.

Inexplicably, when this flag is added, pipx disregards the `--python` flag. During the initial development of the Poetry
installation system, I found that if the path to Python is not specified then pipx will use the system installation of
Python on the GitHub Actions runner machines instead of the installation of the project's standardized version of Python
made via the actions/setup-python action. So now we must switch to using the PIPX_DEFAULT_PYTHON environment variable to
configure which Python installation pipx should use.